third_party/blink/web_tests/media/W3C/video/currentSrc/currentSrc_nonempty_after_setting_src-expected.txt

videoElement.currentSrc should not be empty after setting src attribute (even to a bogus value)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS false is false
PASS successfullyParsed is true

TEST COMPLETE


spec reference
